fix:change score int

+++ database/migrations/2026_04_09_000317_create_calification_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table
             ->foreignId('user_id')
             ->constrained('users');
-            $table->decimal('score', 10, 2);
+            $table->unsignedTinyInteger('score');
             $table->timestamps();
         });
     }
